- Splitt common test code. - Splitt out HandlerTest. - Make JobTest parameterized to run with current way of using frozen bucket concept, or future using BucketExecutor and content layer bucket locking.
