+++ persistence/src/vespa/persistence/dummyimpl/dummy_bucket_executor.cpp
@@ -24,14 +24,14 @@ DummyBucketExecutor::~DummyBucketExecutor() {
 
 std::unique_ptr<BucketTask>
 DummyBucketExecutor::execute(const Bucket & bucket, std::unique_ptr<BucketTask> task) {
-    {
-        std::unique_lock guard(_lock);
-        while (_inFlight.contains(bucket.getBucket())) {
-            _cond.wait(guard);
-        }
-        _inFlight.insert(bucket.getBucket());
-    }
     _executor->execute(makeLambdaTask([this, bucket, bucketTask=std::move(task)]() {
+        {
+            std::unique_lock guard(_lock);
+            while (_inFlight.contains(bucket.getBucket())) {
+                _cond.wait(guard);
+            }
+            _inFlight.insert(bucket.getBucket());
+        }
         bucketTask->run(bucket, makeLambdaCallback([this, bucket]() {
             std::unique_lock guard(_lock);
             assert(_inFlight.contains(bucket.getBucket()));
